requirements.txt
langchain
langchain-community
pypdf
tiktoken
openai
python-dotenv